third_party/blink/web_tests/pointer-lock/pointerlock-then-fullscreen-expected.txt

Test pointerlock then fullscreen.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.testRunner is defined.
     Lock targetDiv1.
PASS onpointerlockchange received after: Lock targetDiv1.
PASS document.pointerLockElement is targetDiv1
PASS document.pointerLockElement is targetDiv1
PASS document.webkitFullscreenElement is targetDiv1
PASS successfullyParsed is true

TEST COMPLETE
